Improve radial bars and add generalization
